[Capture] Misc. improvements/bug fixes to `PlxprInterpreter` (#6913)
This PR adds a few changes to `PlxprInterpreter` to fix some issues that
were found.

* JAX Primitives that accept functions as inputs can accept different
signatures than what shows up in jaxpr equations. The signature is
normalized using the `Primitive.get_bind_params` method. This PR updates
`PlxprInterpreter` to call this before binding primitives. Not really
sure how to test this. Suggestions are welcome.
* Custom registrations for higher-order primitives now have better
handling of constants. If interpreting the inner jaxpr of HOPs
introduces constants, we now correctly handle them.
* As part of this change, `jaxpr_to_jaxpr` no longer returns `Jaxpr`,
but rather `ClosedJaxpr`, so that we can have access to the consts as
well.
* Also fixed a tiny typo in tests.

[sc-83701] [sc-83702]