Update GPU workflow to use the new self-hosted GPU runner infrastructure (#6927)
**Context:**
Swap `tests-gpu.yml` to using the new self-hosted GPU runner
infrastructure.

**Description of the Change:**

The current GPU tests use runner infrastructure that is quite old and
have accumulated a fair amount of technical debt. As part of an overhaul
of the infrastructure, net-new instances have been deployed.

This pull request updates the tests-gpu workflow to use the newly
deployed runners.

**IMPORTANT NOTE** Other than the minor change to `runs-on` the backend
changes should be transparent for users of pennylane. Any and all
behaviors related to gpu workflows should continue working as they are
today.

**Benefits:**

Enable us to use newer infrastructure that have non-deprecated runtimes
(backend).

**Possible Drawbacks:**

It's possible for a test to fail that used to pass on previous gpu
runners. I have taken steps to ensure the build environment between the
outgoing runners and incoming runners have as much parity as possible.

Of course, we will ensure the gpu tests pass before merging this PR, but
maintainers are encouraged to keep an eye on failing gpu tests over the
next couple days to ensure no failures are happening that are
specifically related to build environment changes.

**Related GitHub Issues:**

[sc-83176](https://app.shortcut.com/xanaduai/story/83176/investigate-and-try-updating-lambda-runtime-for-pl-runners-pennylane-lightning)