Fix double QubitUnitary bug from Catalyst (#6926)
**Context:**
A recent deprecation https://github.com/PennyLaneAI/pennylane/pull/6840/
deprecated the usage of `QubitUnitary` type input `base` for
instantiation of `ControlledQubitUnitary`.
However, a modification that was equivalent in PennyLane caused
unexpected behavior in Catalyst,
https://github.com/PennyLaneAI/catalyst/issues/1494.
After investigation, it appears to be the issue with the private method
`_try_wrap_in_custom_ctrl_op` of `pennylane/ops/op_math/controlled.py`,
since it used to call the deprecated `init` to directly use base as
`QubitUnitary`, and to safely deprecate it we decided to locally replace
with equivalent `ControlledQubitUnitary(base=op.matrix(), ...)`, this
caused Catalyst to received double `QubitUnitary`.

It is also noteable that the call `ControlledQubitUnitary` should not
appear in this script, since logically all the classes living in
`controlled_ops` depend on `controlled.py`, which means that from the
very beginning the affected private method served as a tricky hack with
cyclic dependence.

Therefore, we would like to replace the `ControlledQubitUnitary` call
with a direct `ControlledOp`.

**Description of the Change:**
source files: as mentioned above.

tests: necessary improvements were made to correctly reflect the
assertion results.

**Benefits:**
 - Remove bugs
 - Less cyclic dependency

**Possible Drawbacks:**
- Probably there're more that depend on this method, which might be
affected by this fix

**Related GitHub Issues:**
https://github.com/PennyLaneAI/catalyst/issues/1494

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>