`QubitUnitary` accepts sparse matrices as inputs (#6889)
**Context:**
As a follow-up of https://github.com/PennyLaneAI/pennylane/pull/6863, in
this PR we would like to introduce the overloaded `QubitUnitary` capable
of taking sparse matrices and processing along.

Currently, if a user create a new `QubitUnitary` with sparse matrix as
input, it will succeed with `unitary_check=False` but fail with `True`.
This is because lots of the internal processing make use of incompatible
functionalities e.g. `einsum` for `csr_matrix`, for which we will create
more sparse-specific routines to help build a reasonable workflow for
the sparse cases.

**Description of the Change:**
`QubitUnitary` has quite some API that we need to consider, some of
which might need further attention to decide whether or not to implement
for sparse matrices right now.
- [x] `init`: inserted extra conditional unitary check, and separated
the previous logic as static `_unitary_check`
- [x] `matrix()`: the method to generate `QubitUnitary`'s matrix should
be able to return a sparse as well
- [x] `adjoint`: conditional treatment for sparse matrices, since our
default dispatch method does not work for `csr_matrix`
- [x] `pow`: similar as above, although this could be easily merged into
our multi-dispatch in the future
- [x] `compute_decomposition`: to avoid unnecessary extra work, one- and
two-qubit sparse ops are still converted back to desnse matrices,
although in practice it actually does not even make sense to use sparse
matrices at this level (recalling our need for sparse matrices, it's for
systems of size ~20)
- [x] support for wires of size 20: should be able to init a sparse
`QubitUnitary` obj with 20 wires.

**Benefits:**
Necessary extension for `QubitUnitary` implemented.

**Possible Drawbacks:**
Implementations are quite scattered in the source code. In the future we
can think about adding overloading to `qml.math` instead to help unify
the matrix manipulations of sparse matrices; this is needed for
`QubitUnitary` to act like a 'normal' operator as others, or it won't
have enough capabilities, e.g. `QU @ QU` won't work if both implemented
as `csr_matrix`, because PL is confulsed about how to dispatch for CSR's
when it comes to necessary math methods e.g. `reshape`. These are going
to be added in [another
PR](https://github.com/PennyLaneAI/pennylane/pull/6947)

Also, to minimize the amount of conditional branches specifically for
`csr_matrix`, we neglected the following functionalities:
- `_controlled`: this method actually involves the sparse implementation
of `ControlledQubitUnitary`, which is beyond current scope; should be
included in a future story of implementing sparse for generic controlled
ops.
- Any other hidden interfaces that were inherited from parent classes:
same as above; require further discussions and extra stories/epics if
needed. Current interfaces already provide enough features for
`QubitUnitary` to be called in the pipeline. For example, `prod`, `sum`
etc.

P.S. please note that the multiplication between operators and states is
not yet involved here since it is a matter of `apply_operation` which
should be resolved in some follow-ups (e.g.
https://github.com/PennyLaneAI/pennylane/pull/6883#issuecomment-2613345018,
or some other refined smaller PR).

**Related GitHub Issues:**
[sc-82068]

---------

Co-authored-by: Amintor Dusko <87949283+AmintorDusko@users.noreply.github.com>
Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>
Co-authored-by: Isaac De Vlugt <isaacdevlugt@gmail.com>