[Capture] Qfunc jaxpr is cached when executing `QNode`s (#6923)
**Context:**
This PR adds support for caching the jaxpr of a qnode's qfunc.
Additionally, this PR adds support in `QNode` to be able to specify
`static_argnums`.

**Description of the Change:**
* Add `capture_cache` attribute to `QNode`. This is an `LRUCache` that
is supposed to hold the jaxpr for the qfunc for various argument/keyword
argument types.
* Add `static_argnums` argument to `QNode` for tracking `static_argnums`
* Add `_split_static_args` helper function to
`pennylane/workflow/capture_qnode.py` to separate the dynamic and static
arguments of a QNode.
* Add `_get_jaxpr_cache_key` helper function to
`pennylane/workflow/capture_qnode.py` to compute the hash used as a key
for the jaxpr cache. This key depends on the abstract eval (shape +
dtype) of the positional arguments, and the concrete values of the
keyword arguments.
* If `static_argnums` are provided, we use the concrete eval of the
positional arguments to compute the hash.
* If any of the arguments have dynamic shapes, the letter(s)
corresponding to the abstract axis is used in the shape tuple of the
serialized argument.

**Benefits:**
* We can cache jaxpr, eliminating the need to re-compute jaxpr every
time a qnode is executed.

**Possible Drawbacks:**
N/A

**Related GitHub Issues:**

[sc-83774] [sc-84375]