Add class for mid-circuit measurements in a parametrized measurement basis (#6938)
**Context:**
To fully support MBQC-like workloads, and generalising our measurements
and MCMs, we aim to support the use of arbitrary measurement processes,
including ones in arbitrary bases with parametric angles.

**Description of the Change:**

- We add a `ParametrizedMidMeasure` class that can express a MCM in the
XY, ZX or ZY planes, based on plane and angle. It inherits from standard
MCM's and overwrites a few methods to improve labelling and printing, as
well as the adding the methods and attributes needed to track the basis
and diagonalize.
- This MCM will be treated as standard MCM by any check for
`isinstance(op, MidMeasure)`. It must be diagonalized before execution
for any device that doesn't natively support MCMs with an arbitrary
basis, otherwise it will be executed in the computational basis.
- We add a transform to diagonalize these MCMs (currently not
plxpr-compatible, more work to follow).

**Benefits:**
Allowing arbitrary-basis MCMs will allow for direct application of MBQC
like semantics (dynamic corrections using byproduct operators), as well
as opening support for more complex research workloads (e.g. universal
blind quantum computation).

**Possible Drawbacks:**
The diagonalization transform currently modifies MCM attributes
in-place, which I ware generally trying to move away from. However, the
`MeasurementValue` needs to be linked to the `MeasurementProcess` when
created by the `measure` function (next PR, but completely analogous to
what already exists for standard MCMs), so replacing the MCM seems to
get messy, and leaving it as-is seems inaccurate (but maybe
preferable?).


[sc-83334]

---------

Co-authored-by: Joey Carter <joey.snarrcarter@gmail.com>
Co-authored-by: Lee J. O'Riordan <lee@xanadu.ai>
Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>