Fix issue with implicit graph iteration
Wrapping a QubitGraph in a tuple, for example, implicitly iterated over
the underlying qubit graph, which happened since a __getitem__() method
was defined, which gives access to the underlying qubit objects, and is
called in the constructor of a sequential container, like tuple(), when
no __iter__() method is defined. This is fixed by implementing an
QubitGraph __iter__() method that yields itself.