[Capture] `default.qubit` can execute circuits transformed by `qml.defer_measurements` in analytic mode (#6937)
**Context:**
This PR follows up on #6838 to add support to execute circuits
transformed with `defer_measurements` on `default.qubit`. Note that this
PR **does not** integrate the transforms and device preprocessing, so
the transform needs to be manually applied to the program before sending
to the device for execution. That will happen in future work.

**Description of the Change:**
* Update `DefaultQubitInterpreter` to support `ctrl_transform` and
`adjoint_transform` primitives using `CollectOpsandMeas`. This causes
control flow to get unrolled, but I'm not really sure how else to get
them to work.
* Update `DefaultQubitInterpreter.interpret_operation` to renormalize
the state after a `Projector` is applied.
* Added a minor update to
`qml.devices.qubit.measure.get_measurement_function` so that for
`expval`s that don't have an `obs`, we default to using
`state_diagonalizing_gates`. This was needed as terminal MCM
measurements get replaced by measurements on eigvals + wires.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-82054]

---------

Co-authored-by: Astral Cai <astral.cai@xanadu.ai>