add NullMeasurement (#6989)
**Context:**

Performing measurements can consume a non-trivial amount of resources on
large problems. We need to be able to have a dummy measurement process
that allows us to use the full pennylane workflow, without spending any
amount of resources processing a measurement result.

This will allow us to perform benchmarks on large and remote devices
without actually needing to do any additional meausrements.

**Description of the Change:**

Adds a `qml.measurements.NullMeasurement` measurement process that
always corresponds to `np.array(np.nan)`, with a shape `()` and a dtype
`float`. This was chosen so the measurement could be used with jax,
program capture, and catalyst.

**Benefits:**

More informative profiles.

**Possible Drawbacks:**

Don't know if it's useful outside of profiling. Would be nice to have a
`None` result, but that doesn't have a shape and dtype.

**Related GitHub Issues:**

[sc-52374]

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>