[Capture] Switch from binding `qnode_kwargs` to `execution_config` (#6991)
**Context:**

As the capture workflow is getting more complicated, we should start
using the well-defined `ExecutionConfig` object instead of the ambiguous
and unspecified `qnode_kwargs`.

We also need to start passing the `execution_config` to the device, as
that information is needed for handling mid circuit measurements.

**Description of the Change:**

Switches from binding a `qnode_kwargs` dictionary to an
`execution_config` object.

**Benefits:**

Easier to manage the configuration of a workflow. Can specify MCM
configuration info for device execution.

**Possible Drawbacks:**

Technically a breaking change, but a breaking change to an experimental
project.

**Related GitHub Issues:**

[sc-84916]

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>