Add measure functions for MCMs in X, Y and arbitrary measurement basis (#6953)
**Note:** I also added docs rendering in this PR (I meant to in the
previous one), so we should include the docs rendering for the `ftqc`
module in the review process

-----------------

**Context:**
In the existing architecture of PennyLane, if a MCM returns a value that
we would like to manipulate/use following the measurement, the MCM needs
to be stored in a `MeasurementValue`. For the computational basis,
`qml.measure` is the function that creates an MCM, links it to a
`MeasurementValue`, and returns the `MeasurementValue` for use.

We've added a base class for creating MCMs in bases other than the
computational basis to the ftqc module, but it doesn't have this
functionality yet.

**Description of the Change:**
We add `ftqc.measure_arbitrary_basis`, `ftqc.measure_x`, and
`ftqc.measure_y`, analogous to `qml.measure`.
Additionally, we add classes `XMidMeasureMP` and `YMidMeasureMP` (used
in the x and y measurement functions), as subclasses of
`ParametrizedMidMeasureMP`.

**Benefits:**
- the `measure` functions allow us to easily return a measurement value
and use or manipulate is, as in:

```
    m = measure_y(0)
    qml.cond(m, qml.RX)(2.345, 1)
```

- designated classes for X and Y basis measurements will simplify logic
in creating MBQC circuits, which rely mostly on X and Y basis MPs
- designated classes for X and Y basis measurements allows them to be
labelled clearly as X or Y (instead of XY-plane) measurements will
improve circuit visualization


[sc-83347]

---------

Co-authored-by: Joey Carter <joey.snarrcarter@gmail.com>
Co-authored-by: Lee J. O'Riordan <lee@xanadu.ai>
Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: Lee James O'Riordan <mlxd@users.noreply.github.com>