Bump `rng_salt` version to `v0.40.0` (#6854)
**Context:**

With v0.40.0 of Pennylane out, we are safe to bump the `rng_salt`
version.

**Benefits:** 

This alters the RNG salt of our test suite, ensuring that tests from
last release weren't passing due to some magic salt.

**Description of the change:**

This involved modifying some tests but increasing their tolerances. This
was discussed with the core team to ensure we aren't accidently hiding
any problems.

[sc-82812]

---------

Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>