Fix handling of `cumulative_angles` in `merge_rotations` (#7011)
**Context:**

For the `Rot` gate, the angles can cancel in a non-trivial way (e.g.
Rot(φ,0,-φ) = RZ(φ) RY(0) RZ(-φ) = RZ(0) = I).

Unlike `single_qubit_fusion`, this was not being handled correctly and
only the trivial case where all angles are zero was being checked.

**Description of the Change:**

- Create a boolean flag `angles_cancel` that determines whether the
operator is the identity and can be ignored in the transformed tape.
- Added a test to cover the use case.

**Benefits:** Correct behaviour.

**Possible Drawbacks:** None identified.

**Related GitHub Issues:** Fixes
https://github.com/PennyLaneAI/pennylane/issues/7010

[sc-85327]

---------

Co-authored-by: David Wierichs <david.wierichs@xanadu.ai>
Co-authored-by: Christina Lee <christina@xanadu.ai>