Handling `MCM` in captured transforms (#7020)
**Context:** Plxpr transforms such as `cancel_inverses`, which use
dictionaries to hold ops that have been traced, but not yet applied, are
giving incorrect results with MCMs in the circuit.

*This PR updated `single_qubit_fusion` and `cancel_inverses`.*

**Description of the Change:** We apply the previously stored operators
before applying the MCM, so that the result is correct.

**Benefits:** Correct result.

**Possible Drawbacks:** As for other higher-order primitives, we risk
that some optimizations are not applied. For example, in the case of
`cancel inverses`, if we have an operation `O` on wire `0`. a MCM on
wire `1`, and then the inverse of `O` on wire `0`, we are not canceling
`O` with its inverse (although we could, because the measurement is on a
different wire).

This is a further enhancement that can be implemented in the future for
all the higher-order primitives and all the quantum transforms with
program capture enabled.

**Related GitHub Issues:** None.

**Related Shortcut Stories:** [sc-85404]

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>