[Capture] `default.qubit` `defer_measurements` finite shots execution (#6961)
**Context:**
This PR is a follow up on #6838 and #6937 to add finite shot execution
support to `default.qubit`.

**Description of the Change:**
* Update `DefaultQubitInterpreter` to handle the ability to update shots
based on the specified `postselect_mode`. The currently supported
`postselect_mode`s are `"hw-like"` and `"fill-shots"`.
* Update `DeferMeasurementsInterpreter` to not use `aux_wires`. This is
needed so that we don't need to know ahead of time which wires to use to
map MCMs. Instead, now we specify `num_wires`, which should be an
integer for the total number of available wires. When executing, this
should be known since the device requires wires to be specified.
* Because of traceable wires, there might be some sharp bits. Take a
look at the following example:
  ```python
  qml.capture.enable()
  import jax

  @qml.capture.expand_plxpr_transforms
  @partial(qml.defer_measurements, num_wires=1)
  def f(n):
      qml.measure(n)
  ```
  ```pycon
  >>> jax.make_jaxpr(f)(0)
{ lambda ; a:i64[]. let _:AbstractOperator() = CNOT[n_wires=2] a 0 in ()
}
  ```
The jaxpr is correct, but if I try to execute with `n=0`, we will
expectedly fail because we tried to create a `CNOT([0, 0])`. This is
expected, but would be nice if it could be caught at capture time rather
than runtime. I have updated the `defer_measurements` documentation
accordingly. Let me know if there are any other suggestions about how to
deal with this.
* Update `DefaultQubitInterpreter` to accept `ExecutionConfig`.

**Benefits:**

**Possible Drawbacks:**
Using `hw-like` postselection currently will not be jit friendly when
returning `qml.sample`, as the number of samples can be less than the
original number of samples non-deterministically.

**Future Work:**
* Integrate execution with the `mcm_method` and `postselect_mode`
arguments of `QNode`.

**Related GitHub Issues:**

[sc-82064]

---------

Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: albi3ro <chrissie.c.l@gmail.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>