[Capture] Capture the `commute_controlled` transform (#6946)
**Context:** `commute_controlled` can now be captured. That is, it can
be applied when `qml.capture.enabled` is `True`.

**Description of the Change:** As above.

**Benefits:** One more transform can be applied with the new
experimental capture feature.

**Possible Drawbacks:** 

- Right now, the `commute_controlled` function (even in legacy PL) does
not work with abstract wires. The reason is that the `shared_wires`
method of the Wires class can't detect the intersection between abstract
wires. Enabling this feature (if possible) is out-of-scope for this PR,
but a
[story](https://app.shortcut.com/xanaduai/story/85257/commute-controlled-supports-abstract-wires)
has been created to come back to this in the future.

- When the direction is set to `right`, the function is less efficient
than the one in which the direction is `left`. This is because the
interpreter can only scan the equations of the `jaxpr` in one direction
(forward).

**Related GitHub Issues:** None.

[sc-83907]

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>
Co-authored-by: David Wierichs <david.wierichs@xanadu.ai>