Dispatch more `scipy.sparse` functionalities (#6947)
**Context:**
We used to make full use of autoray to automatically deal with most of
the common packages e.g. numpy and torch. Unfortunately, for scipy
especially the scipy.sparse we care about, there's no immediately
feasible way to dispatch functions to the correct places, especially for
the method in `scipy.sparse.linalg`. Therefore this PR is needed if we
want to fully automatically correctly dispatch our math functions to the
scipy sparse backend.

**Description of the Change:**
Single Dispatch
 - [x] det
 - [x] eigs
 - [x] trace
 - [x] reshape
 - [x] real
 - [x] imag
 - [x] sum
 - [x] inv
 - [x] expm
 - [x] matrix_power
 - [x] norm

Multiple Dispatch
 - ~[x] dot~ not needed anymore since the coerce issue resolved

**Benefits:**
Now we can deal with the mostly common linalg compuations with sparse
matrices, without wrongly dispatches to `scipy.linalg`.

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-84132]

---------

Co-authored-by: Amintor Dusko <87949283+AmintorDusko@users.noreply.github.com>
Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>
Co-authored-by: Isaac De Vlugt <isaacdevlugt@gmail.com>
Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>