Add a helper in `math/quantum` to calculate the matrix square root of sparse, hermitian matrices (#6976)
**Context:**
This need arose from https://github.com/PennyLaneAI/pennylane/pull/6963
where we realized that `scipy.sparse` could NOT help with a
sparsity-preserved `sqrtm` as a sparse counterpart of its own
`scipy.linalg.sqrtm` using Schur decomposition. Soon we realized that
there could be a long list of lost functionality, e.g. `eigh`, `sqrtm`
etc. We would like to update step by step such that when we implement
the sparse representation for various matrices we do not have to convert
them into dense repr, but this might involve quite some effort
considering performance benchmark and lack of references. Therefore, we
need some temporary solution that helps with specific `Operator` such as
`BlockEncode` to avoid an explicit cast back to dense, thought the
performance might vary.

For the `BlockEncode` considered in this PR, it specifically used
`sqrtm` for a special class of matrices $Id - A^\dagger A$ where $A$ is
the input matrix to encode into quantum circuits. This is very close to
our own `sqrt_matrix` in `qml.math.quantum` that deals with the square
root computation of a density matrix; our previous dense-matrix
algorithm also made this assumption and directly called `sqrt_matrix` in
the needed place. Hence we put forward a commonly used iterative method
here suitable for similar Hermitian matrices to have square roots.

A quick introduction of what is the Denman–Beavers iteration used by
this PR can be found at
[Wikipedia](https://en.wikipedia.org/wiki/Square_root_of_a_matrix#By_Denman%E2%80%93Beavers_iteration).

**Description of the Change:**
Provide a new helper in `quantum.py`

**Benefits:**
Enable the sparse version of `BlockEncode`

**Possible Drawbacks:**
This method might have poor performance if used improperly. Empirically,
do not go over 100by100 or it will cost at least several seconds to
converge.
It might not converge when the input matrix is away from
positive-definite Hermitian, unlike our existing `sqrt_matrix` that
works for dense matrices, which in general can handle any Hermitian
matrices)

**Related GitHub Issues:**
[sc-84693]

---------

Co-authored-by: David Wierichs <david.wierichs@xanadu.ai>
Co-authored-by: Lee James O'Riordan <mlxd@users.noreply.github.com>