Remove num_diagonalizing_gates from specs (#7047)
**Context:**

`num_diagonalizing_gates` gets exceedingly expensive for observables
that don't have a hard coded diagonalization or belong to subset of
easily diagonalizable situations. Therefore we fallback to exact
diagonalization, which takes forever and consumes massive amounts of
memory. It's not a lightweight, like most other things in `specs`.

It also isn't even very well defined for composite observables. What's
the "diagonalizing gates" for `X(0) + Y(0)`, especially if the device is
going to split it up into two separate executions?

If the device does need to apply the diagonalizing gates, it will apply
`split_non_commuting` and `diagoanlize_measurements` as part of its
preprocessing, and the diagonalizing gates will be absorbed into the
number of operations in the circuit.

**Description of the Change:**

Removes `num_diagonalizing_gates` from `QuantumScript.specs` and
`qml.specs`. This property and function now also return a special
`SpecsDict` that raises more informative errors if the user requests
either `num_diagonalizing_gates` or some key not present in the
dictionary.

**Benefits:**

`specs` is usable for larger problems.

**Possible Drawbacks:**

Someone might have actually been wanting it.

**Related GitHub Issues:**
Fixes #6735 [sc-80996]

---------

Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>