[Capture] Fall back to tape transforms if `plxpr_transform` isn't available (#6922)
**Context:**
For program capture 1.0, we want to support as many transforms as
possible. To do so, this PR aims to support falling back to tape
transforms if a transform does not have an available `plxpr_transform`.

**Description of the Change:**
* Add fallback `plxpr_transform`, which first creates a tape, applies
the transform to it, and then converts the tape back into jaxpr.
* To support transforming `QNode`s correctly, this PR also updated
`CollectOpsandMeas` to have a handler for `QNode`, so that we can now
convert `QNode`s into tapes.

**Benefits:**
* More transforms work with plxpr without custom implementations

**Possible Drawbacks:**
* Transforms that return multiple tapes don't work. We currently raise
an error if transforming a tape returns multiple tapes.
* Transforms with non-trivial post-processing functions don't work. We
don't catch this right now, as catching it would be non-trivial. It is
possible using `ast`. There will likely still be edge cases, but it
_could_ make the feature more robust. That being said, it felt like
overkill to do it, so I did not. But, if reviewers feel like it is
necessary, I can add it. @isaacdevlugt your input for this would be
really helpful.
* Jaxpr that has control flow with dynamic predicates (eg. tracer
condition for `qml.cond`, tracer number of iterations for
`qml.for_loop`, etc.) will not work. I have documented this in the
`qml.transform` docs under the program capture sharp bits section.
@isaacdevlugt this information would also be useful to add to the
program capture sharp bits page.

**Related GitHub Issues:**

[sc-84031] [sc-85652]

---------

Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>