[FTQC] Add data structures and API for qubit mapping (#6962)
**Context:** To provide support for MBQC, and eventual QEC protocols, we
need to maintain a mapping between physical qubits and logical qubits.
This mapping will allow for the expression of operations on logical
qubits that can be mapped to operations on physical qubits at the
hardware layer.

**Description of the Change:** Adds the basic structure for this
mapping, implemented in the `QubitGraph` class, which represents a
hierarchical qubit memory model as nested graphs of qubits. A QubitGraph
is a qubit that contains a graph of underlying qubits, where each
underlying qubit is itself a QubitGraph.

**Example usage:** The example below initializes a qubit as a 3x4
Cartesian grid of "underlying" qubits and uses this qubit in a simple
circuit.

```python
import networkx as nx
import pennylane as qml
from pennylane.ftqc import QubitGraph

graph = nx.grid_2d_graph(3, 4)
q0 = QubitGraph(graph, id=0)

dev = qml.device("default.qubit")

@qml.qnode(dev)
def circuit():
    qml.RX(1.618, wires=q0)
    return qml.expval(qml.PauliZ(q0))
``` 

**Future Work (for subsequent PRs):**

- Consolidate graph-generation functions across the FTQC module.
- Propagate transversal operations on a `QubitGraph` down to its
underlying qubits.
- Enable tensor-like indexing of `QubitGraph` objects. This is already
partially implemented and works for simple, list-like indexing, but more
work is required for more complex indexing. For example, if we have some
number of "data" qubits and some number of "auxiliary" qubits, indexed
as `('data', 0), ('data', 1), ..., ('aux', 0), ('aux', 1), ...`, it may
be useful to select all of the data qubits, for instance, using

    ```python
    q = QubitGraph(...)
    q[("data", :)]
    ```

- [Optional] Add support for both networkx and rustworkx graph objects.

-----


[[sc-84109](https://app.shortcut.com/xanaduai/story/84109/implement-the-qubitgraph-data-structure-in-pennylane)]

---------

Co-authored-by: Lee James O'Riordan <mlxd@users.noreply.github.com>
Co-authored-by: lillian542 <38584660+lillian542@users.noreply.github.com>